datagen/look-ecomm/requirements.txt
faker
pandas
psycopg2-binary
SQLAlchemy
google-cloud-pubsub
